pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Jeroen van Lier <jeroencvlier@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.0"
ujson = "^5.9.0"
tqdm = "^4.66.1"
pyarrow = "^15.0.0"
joblib = "^1.3.2"
matplotlib = "^3.8.2"
scipy = "^1.12.0"
seaborn = "^0.13.2"
scikit-learn = "^1.4.0"
python-dotenv = "^1.0.1"
wandb = "^0.16.3"
statsmodels = "^0.14.1"
fastparquet = "^2024.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ipykernel = "^6.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
